==> code/01_gathering_data.ipynb <==
# Project 3: Web Scraping and NLP: Depression vs Bipolar

## Problem description
​
Provided with numerous posts on Reddit, I had a binary classification problem on hand to see if a difference could be infered between depression and bipolar posts. After scraping two subreddits, I compared Naive Bayes, Logistic Regression, and KNN models to finetune one that would perform the best. My main concern was measuring the accuracy of the model. After, choosing my model, I went ahead and train my model to make real time predictions. In the 'real_time_predictions' subfolder you will find a code that if ran will tell you with some accuracy whether the person who wrote a paragraph about how they feel should be treated for bipolar or depression. 

### Project Structure:
- Notebook 1. Web APIs and Data Collection
- Notebook 2. EDA, Data Cleaning
- Notebook 3. Pre-Processing
- Notebook 4a. Modeling: Naive-Bayes
- Notebook 4b. Modeling: Logistic Regressoin
- Notebook 4c. Modeling: KNN
- Notebook 5. Model Evaluation

## Web Scraping

import requests
import pandas as pd
import numpy as np
import time

#function loops through 500 posts at a time from a subreddit. using the date filter
def get_data(subreddit):
    after_day = 1
    before_day = 0
    length = 0
    final_df = pd.DataFrame()
    while length < 2000:
        url = 'https://api.pushshift.io/reddit/search/submission?subreddit={}&size=500&after={}d&before={}d&sort=asc'.format(subreddit, after_day, before_day)
        response = requests.get(url)
        results = response.json()
        results_df = pd.DataFrame(results['data'])
        
        final_df = pd.concat([final_df, results_df], sort = True)
        
        length = len(final_df['title'].unique())
        after_day += 1
        before_day += 1
        
        time.sleep(5)
    return final_df

bipolar_df = get_data('bipolar') #pulling the bipolar posts

depression_df = get_data('depression') #gathering the depression posts

len(depression_df['title'].unique())

len(bipolar_df['title'].unique())

bipolar_df[['title', 'selftext', 'subreddit']].head()

depression_df[['title', 'selftext', 'subreddit']].head()

bipolar_df.columns

bipolar_df.drop_duplicates(subset = ['title'], inplace = True)

depression_df.drop_duplicates(subset = ['title'], inplace = True)

bipolar_df.shape

depression_df.shape

bipolar_df.to_csv('../data/bipolar_df.csv', index = False)
depression_df.to_csv('../data/depression_df.csv', index = False)
