==> code/02_eda_cleaning.ipynb <==
# Project 3: Web Scraping and NLP: Depression vs Bipolar

## Problem description

Provided with numerous posts on Reddit, I had a binary classification problem on hand to see if a difference could be infered between depression and bipolar posts. After scraping two subreddits, I compared Naive Bayes, Logistic Regression, and KNN models to finetune one that would perform the best. My main concern was measuring the accuracy of the model. After, choosing my model, I went ahead and train my model to make real time predictions. In the 'real_time_predictions' subfolder you will find a code that if ran will tell you with some accuracy whether the person who wrote a paragraph about how they feel should be treated for bipolar or depression. 

### Project Structure:
- Notebook 1. Web APIs and Data Collection
- Notebook 2. EDA, Data Cleaning
- Notebook 3. Pre-Processing
- Notebook 4a. Modeling: Naive-Bayes
- Notebook 4b. Modeling: Logistic Regressoin
- Notebook 4c. Modeling: KNN
- Notebook 5. Model Evaluation

## EDA and Data Cleaning

import pandas as pd
import numpy as np

depression = pd.read_csv('../data/depression_df.csv')
bipolar = pd.read_csv('../data/bipolar_df.csv')

depression.head()

bipolar.head()

#For our modeling and machine learning we will only use information provided under these columns. 
dep_df = depression[['created_utc', 'title', 'selftext', 'subreddit', 'permalink']]
bip_df = bipolar[['created_utc', 'title', 'selftext', 'subreddit', 'permalink']]

dep_df.head()

bip_df.head()

#Here is am concatination the two subreddits together
together_df = pd.concat([dep_df, bip_df], ignore_index=True)
together_df.head()

together_df.tail()

together_df.info()

'''#There are 554 missing text fields. To avoid problems when putting title column and selftext column together
I fill the NaN and the [removed] observations with '**¯\\_(ツ)_/¯**'. Then when I add them togther
I remove the '**¯\\_(ツ)_/¯**'.'''

together_df.isna().sum()

together_df[together_df.selftext.isnull()]

together_df.selftext.fillna('**¯\\_(ツ)_/¯**', inplace = True)
together_df.isna().sum()

together_df.selftext.replace('[removed]', '**¯\\_(ツ)_/¯**', inplace = True)

together_df['title_selftext'] = together_df['title'] + " " + together_df['selftext']
together_df.head()

together_df.loc[1050,'title_selftext']

'''There are page breaks in the data collected marked as \n
            I remove them with .replace'''


together_df['title_selftext'] = [string.replace('\n', '') for string in together_df['title_selftext']]

together_df.loc[1050, 'title_selftext']

together_df.loc[2, 'title_selftext']

together_df['title_selftext'] = [string.replace('**¯\\_(ツ)_/¯**', '') for string in together_df['title_selftext']]

together_df.loc[2, 'title_selftext']

together_df.to_csv('../data/data_cleaned.csv', index = False)
