==> code/03_pre_processing.ipynb <==
# Project 3: Web Scraping and NLP: Depression vs Bipolar

## Problem description

Provided with numerous posts on Reddit, I had a binary classification problem on hand to see if a difference could be infered between depression and bipolar posts. After scraping two subreddits, I compared Naive Bayes, Logistic Regression, and KNN models to finetune one that would perform the best. My main concern was measuring the accuracy of the model. After, choosing my model, I went ahead and train my model to make real time predictions. In the 'real_time_predictions' subfolder you will find a code that if ran will tell you with some accuracy whether the person who wrote a paragraph about how they feel should be treated for bipolar or depression. 

### Project Structure:
- Notebook 1. Web APIs and Data Collection
- Notebook 2. EDA, Data Cleaning
- Notebook 3. Pre-Processing
- Notebook 4a. Modeling: Naive-Bayes
- Notebook 4b. Modeling: Logistic Regressoin
- Notebook 4c. Modeling: KNN
- Notebook 5. Model Evaluation

## Pre-Processing

import pandas as pd
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer
import re
from nltk.corpus import stopwords

df = pd.read_csv('../data/data_cleaned.csv')

df.head()

data_used = df[['subreddit', 'title_selftext']]
data_used.is_copy = None

data_used.head()

#credit to lesson 5.03:
def pre_processing_data(raw_text, words_to_remove):
    letters_only = re.sub("[^a-zA-Z]", " ", raw_text) #pull only words
    tokenizer = RegexpTokenizer(r'\w+')#tokenize
    text_tokens = tokenizer.tokenize(letters_only.lower()) #make everything lower case
    remove_stopwords = [w for w in text_tokens if w not in stopwords.words('english')] #remove english stopwords 
    lemmatizer = WordNetLemmatizer() #instantiate lemmatize
    text_lem = [lemmatizer.lemmatize(i) for i in remove_stopwords] #lemmatize
    words_to_remove = [lemmatizer.lemmatize(i) for i in words_to_remove] #lemmatize the custom words to remove
    words_to_remove = set(words_to_remove) #make the words to remove in to a set
    remaining_words = [w for w in text_lem if w not in words_to_remove] #remove the custom stop words
    return(" ".join(remaining_words))

#these are custom stop words to remove. If we leave them it would be too easy to spot the difference
words_to_remove = ['depression', 'bipolar', 'antidepressant', 'manic', 'mania', 'hypomanic', 'hypomania']

#run our function on our data
data_used['title_selftext'] = [pre_processing_data(string, words_to_remove) for string in data_used['title_selftext']]

data_used.head()

data_used.to_csv('../data/data_pre_processed.csv')
