==> code/04c_modeling_knn.ipynb <==
# Project 3: Web Scraping and NLP: Depression vs Bipolar

## Problem description

Provided with numerous posts on Reddit, I had a binary classification problem on hand to see if a difference could be infered between depression and bipolar posts. After scraping two subreddits, I compared Naive Bayes, Logistic Regression, and KNN models to finetune one that would perform the best. My main concern was measuring the accuracy of the model. After, choosing my model, I went ahead and train my model to make real time predictions. In the 'real_time_predictions' subfolder you will find a code that if ran will tell you with some accuracy whether the person who wrote a paragraph about how they feel should be treated for bipolar or depression.

### Project Structure:
- Notebook 1. Web APIs and Data Collection
- Notebook 2. EDA, Data Cleaning
- Notebook 3. Pre-Processing
- Notebook 4a. Modeling: Naive-Bayes
- Notebook 4b. Modeling: Logistic Regressoin
- Notebook 4c. Modeling: KNN
- Notebook 5. Model Evaluation

## KNN: K Nearest Neighbors

Here I am running KNN both with CountVectorizer and TfidfVectorizer. The model performed relatively well. Not as well as the Naive Bayes, but not as terrible as Logistic Regression. I ran a number of GridSearches while searching for the best parameters. 

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB, GaussianNB
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import GridSearchCV
from sklearn.preprocessing import FunctionTransformer
from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier

df = pd.read_csv('../data/data_pre_processed.csv')

df.dropna(inplace = True)

y = df['subreddit'].map({'depression':0, 'bipolar':1})
y.value_counts(normalize = True)

X = df['title_selftext']
X.head()

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify = y, random_state = 42, test_size = .33)

### KNN: CountVectorizer

pipe_knn = Pipeline([
    ('cvec', CountVectorizer()),
    ('knn', KNeighborsClassifier())
])

pipe_params_knn = {
    'cvec__max_features': [4000, 5000],
    'cvec__min_df': [1, 2],
    'cvec__max_df': [.8, .9],
    'cvec__ngram_range': [(1,1), (1,2)],
    'knn__n_neighbors': [3, 5, 7, 9]
}

gs_knn = GridSearchCV(pipe_knn, pipe_params_knn, cv = 5)

gs_knn.fit(X_train, y_train);

gs_knn.score(X_train, y_train)

gs_knn.score(X_test, y_test)

gs_knn.best_params_

### KNN: TfidfVectorizer

pipe_knn_tfi = Pipeline([
    ('tfidf', TfidfVectorizer()),
    
    ('knn', KNeighborsClassifier())
])

pipe_params_knn_tfi = {
    'tfidf__max_features': [2000, 3000],
    'tfidf__min_df': [1, 2],
    'tfidf__max_df': [.8, .9],
    'tfidf__ngram_range': [(1,1), (1,2)],
    'knn__n_neighbors': [3,5,7,9]
}

gs_knn_tfi = GridSearchCV(pipe_knn_tfi, pipe_params_knn_tfi, cv = 5)

gs_knn_tfi.fit(X_train, y_train);

gs_knn_tfi.score(X_train, y_train)

gs_knn_tfi.score(X_test, y_test)

gs_knn_tfi.best_params_

### CountVectorizer: Search for ultimate K

cvec = CountVectorizer(max_features = 5000, min_df = 2, max_df = .8, ngram_range = (1,2))

X_train_cvec = cvec.fit_transform(X_train)
X_test_cvec = cvec.transform(X_test)


knn_params = {
    'n_neighbors': range(1, 51, 10),
    'p'        : [1, 2],
    'weights'  : ['uniform', 'distance'],
}

knn_gridsearch = GridSearchCV(KNeighborsClassifier(), 
                              knn_params, 
                              cv = 5,
                              verbose = 1)

knn_gridsearch.fit(X_train_cvec, y_train);
print(knn_gridsearch.score(X_train_cvec, y_train))
print(knn_gridsearch.score(X_test_cvec, y_test))

knn_gridsearch.best_params_

### k = 11 vs k = 3

knn_model = KNeighborsClassifier(n_neighbors=11)

knn_model.fit(X_train_cvec, y_train)

knn_model.score(X_train_cvec, y_train)

knn_model.score(X_test_cvec, y_test)

knn_model = KNeighborsClassifier(n_neighbors=3)

knn_model.fit(X_train_cvec, y_train)

knn_model.score(X_train_cvec, y_train)

knn_model.score(X_test_cvec, y_test)
